refactor(heuristics): split extract-fields.ts into per-field modules
- #126: extract-fields.ts (1060 LOC god-module) split into nine
  per-field modules under extract/ (contact, name, education,
  experience, skills, projects, achievements, summary, shared);
  barrel re-export preserves all call-sites, zero snapshot churn.
- #127: docs/segmentation-spike.md — segmentation centralization
  catalog, SectionedResume interface proposal, go/no-go decision,
  and 4 follow-up issues (#132–#135) filed.


Resolves #126
Resolves #127
